+++ server/src/main/java/net/oneandone/stool/server/users/Pair.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright 1&1 Internet AG, https://github.com/1and1/
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package net.oneandone.stool.server.users;
+
+public class Pair {
+    public final Token token;
+    public final User user;
+
+    public Pair(Token token, User user) {
+        this.token = token;
+        this.user = user;
+    }
+}